Revert "Co-Win Notifier v2"
