+++ README.md
@@ -1,0 +1,17 @@
+# Co-Win-Notifier
+
+Read Before Using This Software!
+
+This utitlity is not signed that is provided without a certificate, 
+which may or maynot be flagged as untrusted on some systems.
+
+Read more about this here:
+https://posts.specterops.io/what-is-it-that-makes-a-microsoft-executable-a-microsoft-executable-b43ac612195e
+
+This is a open source program, which does not collect any data about you, 
+and is only able to communicate to the Co-Win servers.
+You can find it on github at:
+https://github.com/CatalystMonish/Co-Win-Notifier
+
+
+
